+++ src/main/java/ma/m3achaba/plantes/controller/AuthController.java
@@ -22,24 +22,25 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/auth")
 public class AuthController {
 
     private final AuthenticationUserService authservice;
     private final UserService userService;
 
-    @PostMapping("/auth/register")
+    @PostMapping("/register")
     public ResponseEntity<RegisterRequest> register(@RequestBody RegisterRequest registerRequest) {
         RegisterRequest registerResponse = authservice.register(registerRequest);
         return new ResponseEntity<>(registerResponse, HttpStatus.CREATED);
     }
 
-    @PostMapping("/auth/login")
+    @PostMapping("/login")
     public ResponseEntity<RegisterRequest> login(@RequestBody RegisterRequest loginRequest) {
         RegisterRequest loginResponse = authservice.login(loginRequest);
         return new ResponseEntity<>(loginResponse, HttpStatus.OK);
     }
 
-    @PostMapping("/auth/refresh")
+    @PostMapping("/refresh")
     public ResponseEntity<RegisterRequest> refreshToken(@RequestBody RegisterRequest refreshTokenRequest) {
         RegisterRequest refreshResponse = authservice.refreshToken(refreshTokenRequest);
         if (refreshResponse != null) {
@@ -48,88 +49,6 @@ public class AuthController {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
-    @GetMapping("/admin/user/{id}")
-    public ResponseEntity<UserResponse> findById(@PathVariable Long id) {
-        return userService.findById(id)
-                .map(ResponseEntity::ok)
-                .orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
-    }
-    @GetMapping("/admin/usersPage")
-    public ResponseEntity<PageResponse<UserResponse>> findAll(
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size,
-            @RequestParam(required = false) String search
-    ) {
-        PageResponse<UserResponse> response = (search != null && !search.isEmpty())
-                ? userService.findAllWithSearch(page, size, search)  // Recherche avec pagination
-                : userService.findAll(page, size);                    // Pagination sans recherche
-        return ResponseEntity.ok(response);
-    }
-    @GetMapping("/admin/users")
-    public ResponseEntity<List<UserResponse>> findAllUsers(
-            @RequestParam(required = false) String search
-    ) {
-        List<UserResponse> responses = (search != null && !search.isEmpty())
-                ? userService.findAllWithSearch(search)
-                : userService.findAll();
-        return ResponseEntity.ok(responses);
-    }
-    @PostMapping("/admin/addUser")
-    public UserResponse saveUser(
-            @Validated(OnCreate.class) @RequestBody UserRequest request
-    ) {
-        return userService.save(request)
-                .orElseThrow(() -> new ResourceNotFoundException("Unable to save User"));
-    }
 
-    @PutMapping("/admin/updateUser{id}")
-    public UserResponse updateUser(
-            @PathVariable Long id,
-            @Validated(OnUpdate.class) @RequestBody UserRequest request
-    ) {
-        return userService.update(request, id)
-                .orElseThrow(() -> new ResourceNotFoundException("Unable to update User with id: " + id));
-    }
-
-    @DeleteMapping("admin/deleteUser/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Long id) {
-        userService.delete(id);
-    }
-
-    @GetMapping("/user/current")
-    public ResponseEntity<User> getCurrentUser() {
-        try {
-            User userResponse = userService.getCurrentUser();
-            return ResponseEntity.ok(userResponse);
-        } catch (IllegalStateException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
-        }
-    }
-
-    @GetMapping("/user/profile")
-    public ResponseEntity<RegisterRequest> getProfile() {
-        try {
-            RegisterRequest profile = userService.getProfile();
-            return ResponseEntity.ok(profile);
-        } catch (IllegalStateException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
-        }
-    }
-
-    @PutMapping("/user/updateProfile")
-    public ResponseEntity<RegisterRequest> updateProfile(
-            @Validated(OnUpdate.class) @RequestBody RegisterRequest request) {
-        try {
-            RegisterRequest updatedProfile = userService.updateProfile(request);
-            return ResponseEntity.ok(updatedProfile);
-        } catch (IllegalStateException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body(RegisterRequest.builder()
-                            .message("Erreur de mise à jour du profil : " + e.getMessage())
-                            .statusCode(400)
-                            .build());
-        }
-    }
 }
